pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "ct_scan_mlops"
version = "0.0.1"
description = "Chest CT scan multi-classification model for lung tumor detection using PyTorch Lightning"
authors = [
  { name = "MLOps Group 2", email = "s204696@dtu.dk, s250898@dtu.dk" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps", "medical imaging", "chest CT", "lung cancer"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Deep Learning
    "torch>=2.7.0",
    "torchvision>=0.16.0",
    "pytorch-lightning>=2.1.0",
    "torchmetrics>=1.0.0",            # Metrics for training
    "rich>=13.0.0",                   # Rich progress bars
    "timm>=0.9.0",                    # Pretrained models (ResNet18)
    # Image Processing & Augmentation
    "albumentations>=1.3.0",
    "pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",
    # MLOps Tools
    "wandb[workspaces]>=0.24.0",                  # Experiment tracking
    "hydra-core>=1.3.0",              # Configuration management
    "omegaconf>=2.3.0",
    "dvc>=3.0.0",                     # Data version control
    "dvc-gs>=3.0.0",                  # GCS support for DVC
    "loguru>=0.7.3",                  # Application logging
    # API & CLI
    "fastapi>=0.115.0",
    "typer>=0.9.0",
    "uvicorn>=0.30.0",
    # Data Science
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pandas>=2.0.0",
    "kagglehub>=0.4.0",
    "torch-tb-profiler>=0.4.3",
    "python-multipart>=0.0.21",
    "httpx>=0.28.1",
    "onnx>=1.20.1",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.7",
    "streamlit>=1.53.0",
    "psutil>=6.0.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "evidently>=0.7.20",
    "prometheus-client>=0.24.1",
    # Radiomics Feature Extraction
    "scikit-image>=0.21.0",         # GLCM, regionprops
    "PyWavelets>=1.4.0",            # Wavelet features
    "joblib>=1.3.0",                # Parallel processing
    "scipy>=1.16.3",
]

[dependency-groups]
dev = [
    "coverage>=7.6.0",
    "invoke>=2.2.0",
    "mkdocs-material>=9.4.0",
    "mkdocs>=1.6.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
    "mkdocstrings-python>=1.12.0",
    "locust>=2.20.0",                 # Load testing
    "mypy>=1.19.1",
    "pip-audit>=2.10.0",
    "bandit>=1.9.3",
]


# Platform-specific PyTorch configuration
# Linux: CUDA 12.8 (latest stable)
# Windows: CUDA 12.8 (for teammates with NVIDIA GPUs)
# macOS: CPU-only (no CUDA support on Mac)

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
    { index = "pytorch-cu128", marker = "sys_platform == 'win32'" },
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
    { index = "pytorch-cu128", marker = "sys_platform == 'win32'" },
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes (undefined vars, unused imports)
    "I",    # isort (import sorting)
    "UP",   # pyupgrade (modern Python syntax)
    "B",    # flake8-bugbear (common bugs/design problems)
    "SIM",  # flake8-simplify (simplify code)
    "PTH",  # flake8-use-pathlib (use pathlib over os.path)
    "NPY",  # NumPy-specific rules (deprecated APIs)
    "C4",   # flake8-comprehensions (better comprehensions)
    "RET",  # flake8-return (good return practices)
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["ct_scan_mlops"]
known-third-party = ["wandb", "torch", "lightning"]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["D100", "D103"]
"reports/*.py" = ["E", "W", "F", "I", "UP", "B", "SIM", "PTH", "NPY", "C4", "RET"]  # External template, skip linting

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "requires_data: mark test as requiring actual data files (skipped in CI)",
    "slow: mark test as slow running",
]
# Run tests without data by default (CI-friendly)
addopts = "-v --tb=short"

[tool.mypy]
ignore_missing_imports = true
exclude = ["reports/"]

[[tool.mypy.overrides]]
module = ["wandb", "wandb.*"]
follow_imports = "skip"

[tool.coverage.run]
omit = [
    "tests/*",
    "*/config*.py",  # Hydra temporary config files
    ".venv/*",
]

[project.scripts]
train = "ct_scan_mlops.train:train"
sweep-train = "ct_scan_mlops.sweep_train:main"
evaluate = "ct_scan_mlops.evaluate:evaluate"
visualize = "ct_scan_mlops.visualize:visualize"
api = "ct_scan_mlops.api:app"
